+++ lib/routemaster.dart
@@ -18,6 +18,7 @@ import 'src/route_data.dart';
 part 'src/pages/page_stack.dart';
 part 'src/pages/tab_pages.dart';
 part 'src/pages/basic_pages.dart';
+part 'src/observers.dart';
 
 typedef RoutemasterBuilder = Widget Function(
   BuildContext context,
@@ -36,7 +37,7 @@ typedef UnknownRouteCallback = Page Function(
 class DefaultUnknownRoutePage extends StatelessWidget {
   final String path;
 
-  DefaultUnknownRoutePage({required this.path});
+  const DefaultUnknownRoutePage({required this.path});
 
   @override
   Widget build(BuildContext context) {
@@ -47,7 +48,10 @@ class DefaultUnknownRoutePage extends StatelessWidget {
 }
 
 /// An abstract class that can provide a map of routes
+@immutable
 abstract class RouteConfig {
+  const RouteConfig();
+
   /// Called when there's no match for a route. By default this returns
   /// [DefaultUnknownRoutePage], a simple page not found page.
   ///
@@ -93,6 +97,7 @@ abstract class DefaultRouterConfig extends RouteConfig {
 }
 
 /// A standard simple routing table which takes a map of routes.
+@immutable
 class RouteMap extends DefaultRouterConfig {
   /// A map of paths and [PageBuilder] delegates that return [Page] objects to
   /// build.
@@ -196,10 +201,12 @@ class RoutemasterDelegate extends RouterDelegate<RouteData>
   bool _isBuilding = false;
   bool _isDisposed = false;
   late BuildContext _context;
+  final List<RoutemasterObserver> observers;
 
   RoutemasterDelegate({
     required this.routesBuilder,
     this.transitionDelegate,
+    this.observers = const [],
   }) {
     _state.routemaster._delegate = this;
   }
@@ -325,12 +332,20 @@ class RoutemasterDelegate extends RouterDelegate<RouteData>
       return;
     }
 
-    final routeData = _state.stack!._getCurrentPages().last.routeData;
-    print("Updated path: '${routeData.path}'");
-    _state.currentConfiguration = RouteData(
-      routeData.path,
-      isReplacement: routeData.isReplacement,
-    );
+    final pageWrapper = _state.stack!._getCurrentPages().last;
+    final routeData = pageWrapper.routeData;
+
+    if (_state.currentConfiguration!.path != routeData.path) {
+      _state.currentConfiguration = RouteData(
+        routeData.path,
+        isReplacement: routeData.isReplacement,
+        pathTemplate: routeData.pathTemplate,
+      );
+
+      for (final observer in observers) {
+        observer.didChangeRoute(routeData, pageWrapper._getOrCreatePage());
+      }
+    }
   }
 
   // Called when a new URL is set. The RouteInformationParser will parse the
@@ -349,7 +364,7 @@ class RoutemasterDelegate extends RouterDelegate<RouteData>
   Future<void> setInitialRoutePath(RouteData configuration) {
     assert(!_isDisposed);
 
-    _state.currentConfiguration = RouteData(configuration.path);
+    _state.currentConfiguration = configuration;
     return SynchronousFuture(null);
   }
 
@@ -554,21 +569,17 @@ class RoutemasterDelegate extends RouterDelegate<RouteData>
     required RouteData routeData,
   }) {
     while (page is ProxyPage) {
-      if (page is GuardedPage) {
-        if (!page.validate(routeData, _context)) {
-          print("Validation failed for '${routeData.path}'");
-
-          if (page.onValidationFailed == null) {
-            return _NotFoundResult();
-          }
-
-          final result = page.onValidationFailed!(routeData, _context);
-          return _createPageWrapper(
-            routeRequest: routeRequest,
-            page: result,
-            routeData: routeData,
-          );
+      if (page is GuardedPage && !page.validate(routeData, _context)) {
+        if (page.onValidationFailed == null) {
+          return _NotFoundResult();
         }
+
+        final result = page.onValidationFailed!(routeData, _context);
+        return _createPageWrapper(
+          routeRequest: routeRequest,
+          page: result,
+          routeData: routeData,
+        );
       }
 
       page = page.child;
@@ -593,8 +604,6 @@ class RoutemasterDelegate extends RouterDelegate<RouteData>
 
   List<PageWrapper> _onUnknownRoute(_RouteRequest routeRequest) {
     final requestedPath = routeRequest.path;
-    print("Router couldn't find a match for path '$requestedPath'");
-
     final result = _state.routeConfig!.onUnknownRoute(
       requestedPath,
       _context,
@@ -633,6 +642,7 @@ class RoutemasterDelegate extends RouterDelegate<RouteData>
   }
 }
 
+@immutable
 abstract class _PageResult {}
 
 class _PageWrapperResult extends _PageResult {
@@ -778,11 +788,13 @@ class _RouteRequest {
 class StackNavigator extends StatefulWidget {
   final PageStack stack;
   final TransitionDelegate transitionDelegate;
+  final List<NavigatorObserver> observers;
 
-  StackNavigator({
+  const StackNavigator({
     Key? key,
     required this.stack,
     this.transitionDelegate = const DefaultTransitionDelegate<dynamic>(),
+    this.observers = const [],
   }) : super(key: key);
 
   @override
@@ -798,14 +810,19 @@ class StackNavigator extends StatefulWidget {
 class StackNavigatorState extends State<StackNavigator> {
   final _navigatorKey = GlobalKey<NavigatorState>(debugLabel: 'PageStack');
   late Navigator _navigator;
+  late Routemaster _routemaster;
   final HeroController _heroController =
       MaterialApp.createMaterialHeroController();
+  late _RelayingNavigatorObserver _proxyNavigationObserver;
 
   @override
   void initState() {
     super.initState();
 
     widget.stack.addListener(_onStackChanged);
+    _proxyNavigationObserver = _RelayingNavigatorObserver(
+      () => _routemaster._delegate.observers,
+    );
     _updateNavigator();
   }
 
@@ -818,12 +835,22 @@ class StackNavigatorState extends State<StackNavigator> {
       widget.stack.addListener(_onStackChanged);
       _updateNavigator();
     }
+
+    if (oldWidget.observers != widget.observers) {
+      _updateNavigator();
+    }
   }
 
   @override
   void dispose() {
     widget.stack.removeListener(_onStackChanged);
     super.dispose();
+  }
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    _routemaster = Routemaster.of(this.context);
   }
 
   void _onStackChanged() {
@@ -833,7 +860,7 @@ class StackNavigatorState extends State<StackNavigator> {
   }
 
   void _updateDelegate() {
-    Routemaster.of(this.context)._delegate._markNeedsUpdate();
+    _routemaster._delegate._markNeedsUpdate();
   }
 
   void _updateNavigator() {
@@ -848,6 +875,7 @@ class StackNavigatorState extends State<StackNavigator> {
       },
       transitionDelegate: widget.transitionDelegate,
       pages: widget.stack.createPages(),
+      observers: [_proxyNavigationObserver, ...widget.observers],
     );
 
     widget.stack._attachedNavigatorKey = _navigatorKey;