+++ lib/src/pages/basic_pages.dart
@@ -2,6 +2,8 @@ part of '../../routemaster.dart';
 
 /// A page that can create a state.
 abstract class StatefulPage<T> extends Page<T> {
+  const StatefulPage();
+
   PageState createState(Routemaster routemaster, RouteData info);
 
   @override
@@ -29,7 +31,14 @@ abstract class PageWrapper {
   bool maybeSetChildPages(Iterable<PageWrapper> pages);
 
   /// Gets the actual Flutter [Page] object for passing to a [Navigator].
+  ///
+  /// This will only be called once per [PageWrapper], and the result cached.
   Page createPage();
+
+  Page? _page;
+  Page _getOrCreatePage() {
+    return _page ??= createPage();
+  }
 }
 
 /// A page's state, similar to [State] for a [StatefulWidget]. For instance,
@@ -42,7 +51,7 @@ abstract class PageState extends PageWrapper {}
 abstract class ProxyPage<T> extends Page<T> {
   final Page<T> child;
 
-  ProxyPage({required this.child});
+  const ProxyPage({required this.child});
 
   @override
   Route<T> createRoute(BuildContext context) {